Move search bar components to their own subfolder SearchBar for organisation
